move models into separate file
